+++ packages/grid/src/features/standard-table/config/StandardTableConfig.ts
@@ -159,17 +159,7 @@ export interface StandardTableConfig<
   /**
    * Enable sticky behaviour for header and row checkbox
    */
-  stickyHeaderAndRowCheckbox?: boolean;
-
-  /**
-   * Enable a shadow on the right side of the header checkbox
-   */
-  stickyHeaderCheckboxRightShadow?: boolean;
-
-  /**
-   * Enable a shadow on the right side of the row checkbox
-   */
-  stickyRowCheckboxRightShadow?: boolean;
+  stickyCheckboxColumn?: boolean;
 
   /**
    * Set a custom z index